working version: split decompression function in separate file, implement echo and zlib decompressor, add first unittest for it. still needs cleanup
