+++ cvmfs/CMakeLists.txt
@@ -60,10 +60,14 @@ if (BUILD_CVMFS OR BUILD_LIBCVMFS)
        compression/compression.cc
        compression/compressor_echo.cc
        compression/compressor_zlib.cc
+       compression/decompression.cc
+       compression/decompressor_echo.cc
+       compression/decompressor_zlib.cc
        compression/input_cache.cc
        compression/input_file.cc
        compression/input_mem.cc
        compression/input_path.cc
+       compression/util.cc
        directory_entry.cc
        duplex_fuse.cc
        fd_refcount_mgr.cc
@@ -145,10 +149,14 @@ if (BUILD_CVMFS)
                   compression/compression.cc
                   compression/compressor_echo.cc
                   compression/compressor_zlib.cc
+                  compression/decompression.cc
+                  compression/decompressor_echo.cc
+                  compression/decompressor_zlib.cc
                   compression/input_cache.cc
                   compression/input_file.cc
                   compression/input_mem.cc
                   compression/input_path.cc
+                  compression/util.cc
                   cvmfs_fsck.cc
                   statistics.cc
   )
@@ -510,10 +518,14 @@ if (BUILD_LIBCVMFS_CACHE)
                   compression/compression.cc
                   compression/compressor_echo.cc
                   compression/compressor_zlib.cc
+                  compression/decompression.cc
+                  compression/decompressor_echo.cc
+                  compression/decompressor_zlib.cc
                   compression/input_cache.cc
                   compression/input_file.cc
                   compression/input_mem.cc
                   compression/input_path.cc
+                  compression/util.cc
                   fd_refcount_mgr.cc
                   manifest.cc
                   quota.cc
@@ -594,10 +606,14 @@ if (BUILD_SERVER)
        compression/compression.cc
        compression/compressor_echo.cc
        compression/compressor_zlib.cc
+       compression/decompression.cc
+       compression/decompressor_echo.cc
+       compression/decompressor_zlib.cc
        compression/input_cache.cc
        compression/input_file.cc
        compression/input_mem.cc
        compression/input_path.cc
+       compression/util.cc
        directory_entry.cc
        file_chunk.cc
        gateway_util.cc
@@ -744,10 +760,14 @@ if (BUILD_SERVER)
        compression/compression.cc
        compression/compressor_echo.cc
        compression/compressor_zlib.cc
+       compression/decompression.cc
+       compression/decompressor_echo.cc
+       compression/decompressor_zlib.cc
        compression/input_cache.cc
        compression/input_file.cc
        compression/input_mem.cc
        compression/input_path.cc
+       compression/util.cc
        directory_entry.cc
        gateway_util.cc
        globals.cc
@@ -860,10 +880,14 @@ if (BUILD_SERVER)
        compression/compression.cc
        compression/compressor_echo.cc
        compression/compressor_zlib.cc
+       compression/decompression.cc
+       compression/decompressor_echo.cc
+       compression/decompressor_zlib.cc
        compression/input_cache.cc
        compression/input_file.cc
        compression/input_mem.cc
        compression/input_path.cc
+       compression/util.cc
        directory_entry.cc
        manifest.cc
        network/dns.cc
@@ -1048,10 +1072,14 @@ if(BUILD_RECEIVER)
        compression/compression.cc
        compression/compressor_echo.cc
        compression/compressor_zlib.cc
+       compression/decompression.cc
+       compression/decompressor_echo.cc
+       compression/decompressor_zlib.cc
        compression/input_cache.cc
        compression/input_file.cc
        compression/input_mem.cc
        compression/input_path.cc
+       compression/util.cc
        directory_entry.cc
        gateway_util.cc
        globals.cc
@@ -1216,10 +1244,14 @@ if (BUILD_PRELOADER)
                   compression/compression.cc
                   compression/compressor_echo.cc
                   compression/compressor_zlib.cc
+                  compression/decompression.cc
+                  compression/decompressor_echo.cc
+                  compression/decompressor_zlib.cc
                   compression/input_cache.cc
                   compression/input_file.cc
                   compression/input_mem.cc
                   compression/input_path.cc
+                  compression/util.cc
                   gateway_util.cc
                   globals.cc
                   history_sql.cc
