+++ test/micro-benchmarks/CMakeLists.txt
@@ -21,10 +21,14 @@ set (CVMFS_UBENCHMARKS_SOURCES
   ${CVMFS_SOURCE_DIR}/compression/compression.cc
   ${CVMFS_SOURCE_DIR}/compression/compressor_echo.cc
   ${CVMFS_SOURCE_DIR}/compression/compressor_zlib.cc
+  ${CVMFS_SOURCE_DIR}/compression/decompression.cc
+  ${CVMFS_SOURCE_DIR}/compression/decompressor_echo.cc
+  ${CVMFS_SOURCE_DIR}/compression/decompressor_zlib.cc
   ${CVMFS_SOURCE_DIR}/compression/input_cache.cc
   ${CVMFS_SOURCE_DIR}/compression/input_file.cc
   ${CVMFS_SOURCE_DIR}/compression/input_mem.cc
   ${CVMFS_SOURCE_DIR}/compression/input_path.cc
+  ${CVMFS_SOURCE_DIR}/compression/util.cc
   ${CVMFS_SOURCE_DIR}/crypto/hash.cc
   ${CVMFS_SOURCE_DIR}/directory_entry.cc
   ${CVMFS_SOURCE_DIR}/glue_buffer.cc
