+++ test/quickcheck/CMakeLists.txt
@@ -41,10 +41,14 @@ set (CVMFS_QC_SOURCES
   ${CVMFS_SOURCE_DIR}/compression/compression.cc
   ${CVMFS_SOURCE_DIR}/compression/compressor_echo.cc
   ${CVMFS_SOURCE_DIR}/compression/compressor_zlib.cc
+  ${CVMFS_SOURCE_DIR}/compression/decompression.cc
+  ${CVMFS_SOURCE_DIR}/compression/decompressor_echo.cc
+  ${CVMFS_SOURCE_DIR}/compression/decompressor_zlib.cc
   ${CVMFS_SOURCE_DIR}/compression/input_cache.cc
   ${CVMFS_SOURCE_DIR}/compression/input_file.cc
   ${CVMFS_SOURCE_DIR}/compression/input_mem.cc
   ${CVMFS_SOURCE_DIR}/compression/input_path.cc
+  ${CVMFS_SOURCE_DIR}/compression/util.cc
   ${CVMFS_SOURCE_DIR}/crypto/hash.cc
   ${CVMFS_SOURCE_DIR}/crypto/signature.cc
   ${CVMFS_SOURCE_DIR}/directory_entry.cc
