+++ test/stress/CMakeLists.txt
@@ -5,10 +5,14 @@ add_executable(s3benchmark
                ${CVMFS_SOURCE_DIR}/compression/compression.cc
                ${CVMFS_SOURCE_DIR}/compression/compressor_echo.cc
                ${CVMFS_SOURCE_DIR}/compression/compressor_zlib.cc
+               ${CVMFS_SOURCE_DIR}/compression/decompression.cc
+               ${CVMFS_SOURCE_DIR}/compression/decompressor_echo.cc
+               ${CVMFS_SOURCE_DIR}/compression/decompressor_zlib.cc
                ${CVMFS_SOURCE_DIR}/compression/input_cache.cc
                ${CVMFS_SOURCE_DIR}/compression/input_file.cc
                ${CVMFS_SOURCE_DIR}/compression/input_mem.cc
                ${CVMFS_SOURCE_DIR}/compression/input_path.cc
+               ${CVMFS_SOURCE_DIR}/compression/util.cc
                ${CVMFS_SOURCE_DIR}/gateway_util.cc
                ${CVMFS_SOURCE_DIR}/json_document.cc
                ${CVMFS_SOURCE_DIR}/network/dns.cc
