+++ test/unittests/CMakeLists.txt
@@ -41,10 +41,14 @@ if (BUILD_LIBCVMFS_CACHE)
                   ${CVMFS_SOURCE_DIR}/compression/compression.cc
                   ${CVMFS_SOURCE_DIR}/compression/compressor_echo.cc
                   ${CVMFS_SOURCE_DIR}/compression/compressor_zlib.cc
+                  ${CVMFS_SOURCE_DIR}/compression/decompression.cc
+                  ${CVMFS_SOURCE_DIR}/compression/decompressor_echo.cc
+                  ${CVMFS_SOURCE_DIR}/compression/decompressor_zlib.cc
                   ${CVMFS_SOURCE_DIR}/compression/input_cache.cc
                   ${CVMFS_SOURCE_DIR}/compression/input_file.cc
                   ${CVMFS_SOURCE_DIR}/compression/input_mem.cc
                   ${CVMFS_SOURCE_DIR}/compression/input_path.cc
+                  ${CVMFS_SOURCE_DIR}/compression/util.cc
                   ${CVMFS_SOURCE_DIR}/manifest.cc
                   ${CVMFS_SOURCE_DIR}/quota.cc
   )
@@ -166,10 +170,14 @@ if (BUILD_SERVER)
        ${CVMFS_SOURCE_DIR}/compression/compression.cc
        ${CVMFS_SOURCE_DIR}/compression/compressor_echo.cc
        ${CVMFS_SOURCE_DIR}/compression/compressor_zlib.cc
+       ${CVMFS_SOURCE_DIR}/compression/decompression.cc
+       ${CVMFS_SOURCE_DIR}/compression/decompressor_echo.cc
+       ${CVMFS_SOURCE_DIR}/compression/decompressor_zlib.cc
        ${CVMFS_SOURCE_DIR}/compression/input_cache.cc
        ${CVMFS_SOURCE_DIR}/compression/input_file.cc
        ${CVMFS_SOURCE_DIR}/compression/input_mem.cc
        ${CVMFS_SOURCE_DIR}/compression/input_path.cc
+       ${CVMFS_SOURCE_DIR}/compression/util.cc
        ${CVMFS_SOURCE_DIR}/directory_entry.cc
        ${CVMFS_SOURCE_DIR}/gateway_util.cc
        ${CVMFS_SOURCE_DIR}/globals.cc
@@ -436,10 +444,14 @@ set(CVMFS_UNITTEST_SOURCES
   ${CVMFS_SOURCE_DIR}/compression/compression.cc
   ${CVMFS_SOURCE_DIR}/compression/compressor_echo.cc
   ${CVMFS_SOURCE_DIR}/compression/compressor_zlib.cc
+  ${CVMFS_SOURCE_DIR}/compression/decompression.cc
+  ${CVMFS_SOURCE_DIR}/compression/decompressor_echo.cc
+  ${CVMFS_SOURCE_DIR}/compression/decompressor_zlib.cc
   ${CVMFS_SOURCE_DIR}/compression/input_cache.cc
   ${CVMFS_SOURCE_DIR}/compression/input_file.cc
   ${CVMFS_SOURCE_DIR}/compression/input_mem.cc
   ${CVMFS_SOURCE_DIR}/compression/input_path.cc
+  ${CVMFS_SOURCE_DIR}/compression/util.cc
   ${CVMFS_SOURCE_DIR}/cvmfs_suid_util.cc
   ${CVMFS_SOURCE_DIR}/directory_entry.cc
   ${CVMFS_SOURCE_DIR}/duplex_fuse.cc
